refactor: moved classes to different files
